Updated Two Files
InstallMe - Updated script, however still have an issue with the paramaters. Need to double check that it works as expected.

Created a new file for Installing Vigrant, it checks the Hashicorp website for any download url's that end with .exe or .msi, it'll then split the URL into parts and then find the first instance of .msi and name it executableLink, it then creates a variable to the path of the download and then installs the software.

The next task is to have two options install and uninstall.